feat: JPA 적용
